+++ great_expectations/render/renderer/v3/suite_profile_notebook_renderer.py
@@ -102,7 +102,7 @@ checkpoint_config = {
     ]
 }
 checkpoint = SimpleCheckpoint(
-    f"{validator.batch_cache.active_batch_definition.data_asset_name}_{expectation_suite_name}",
+    f"{validator.active_batch.batch_definition.data_asset_name}_{expectation_suite_name}",
     context,
     **checkpoint_config
 )
