+++ great_expectations/render/v3/notebook_assets/suite_edit/footer.py.j2
@@ -12,7 +12,7 @@ checkpoint_config = {
     ]
 }
 checkpoint = SimpleCheckpoint(
-    f"{validator.batch_cache.active_batch_definition.data_asset_name}_{expectation_suite_name}",
+    f"{validator.active_batch.batch_definition.data_asset_name}_{expectation_suite_name}",
     context,
     **checkpoint_config
 )
