+++ great_expectations/validator/validator.py
@@ -8,11 +8,10 @@ import traceback
 import warnings
 from collections import defaultdict, namedtuple
 from collections.abc import Hashable
-from typing import Any, Callable, Dict, Iterable, List, Optional, Set, Tuple, Union
+from typing import Any, Callable, Dict, List, Optional, Set, Tuple, Union
 
 from dateutil.parser import parse
 from marshmallow import ValidationError
-from tqdm.auto import tqdm
 
 from great_expectations import __version__ as ge_version
 from great_expectations.core.batch import Batch
@@ -35,18 +34,11 @@ from great_expectations.dataset.sqlalchemy_dataset import SqlAlchemyBatchReferen
 from great_expectations.exceptions import (
     GreatExpectationsError,
     InvalidExpectationConfigurationError,
-    MetricResolutionError,
 )
-from great_expectations.execution_engine import (
-    ExecutionEngine,
-    PandasExecutionEngine,
-    SparkDFExecutionEngine,
-    SqlAlchemyExecutionEngine,
-)
+from great_expectations.execution_engine import ExecutionEngine
 from great_expectations.execution_engine.pandas_batch_data import PandasBatchData
 from great_expectations.expectations.registry import (
     get_expectation_impl,
-    get_metric_provider,
     list_registered_expectation_implementations,
 )
 from great_expectations.rule_based_profiler import RuleBasedProfilerResult
@@ -70,6 +62,9 @@ from great_expectations.rule_based_profiler.rule_based_profiler import (
 from great_expectations.types import ClassConfig
 from great_expectations.util import load_class, verify_dynamic_loading_support
 from great_expectations.validator.exception_info import ExceptionInfo
+from great_expectations.validator.metric_computation_handler import (
+    MetricComputationHandler,
+)
 from great_expectations.validator.metric_configuration import MetricConfiguration
 from great_expectations.validator.validation_graph import (
     ExpectationValidationGraph,
@@ -88,9 +83,6 @@ except ImportError:
     logger.debug(
         "Unable to load pandas; install optional pandas dependency for support."
     )
-
-MAX_METRIC_COMPUTATION_RETRIES: int = 3
-
 
 ValidationStatistics = namedtuple(
     "ValidationStatistics",
@@ -149,7 +141,10 @@ class Validator:
         self._execution_engine = execution_engine
         self._expose_dataframe_methods = False
 
-        self._show_progress_bars = self._determine_progress_bars()
+        self._metric_computation_handler = MetricComputationHandler(
+            execution_engine=execution_engine,
+            show_progress_bars=self._determine_progress_bars(),
+        )
 
         self.interactive_evaluation = interactive_evaluation
         self._initialize_expectations(
@@ -170,6 +165,104 @@ class Validator:
             self.set_default_expectation_argument("include_config", True)
 
         self._include_rendered_content = include_rendered_content
+
+    @property
+    def execution_engine(self) -> ExecutionEngine:
+        """Returns the execution engine being used by the validator at the given time"""
+        return self._execution_engine
+
+    @property
+    def metric_computation_handler(self) -> MetricComputationHandler:
+        """Returns the "MetricComputationHandler" object being used by the Validator to handle metrics computations."""
+        return self._metric_computation_handler
+
+    @property
+    def batch_cache(self) -> BatchCache:
+        """Returns the "BatchCache" object being used by the Validator to store "Batch" objects in use."""
+        return self._batch_cache
+
+    @property
+    def batches(self) -> Dict[str, Batch]:
+        """Convenience property that returns ordered dictionary of "Batch" objects in use (with batch_id as key)."""
+        return self._batch_cache.batches
+
+    @property
+    def active_batch(self) -> Batch:
+        """Convenience property that returns most recent ("active") "Batch" objects in use."""
+        return self._batch_cache.active_batch
+
+    @property
+    def data_context(self) -> Optional["DataContext"]:  # noqa: F821
+        """Reference to DataContext object handle."""
+        return self._data_context
+
+    @property
+    def expose_dataframe_methods(self) -> bool:
+        """The "expose_dataframe_methods" getter property."""
+        return self._expose_dataframe_methods
+
+    @expose_dataframe_methods.setter
+    def expose_dataframe_methods(self, value: bool) -> None:
+        """The "expose_dataframe_methods" setter property."""
+        self._expose_dataframe_methods = value
+
+    def get_metric(
+        self,
+        metric: MetricConfiguration,
+    ) -> Any:
+        """Convenience method that returns the value of the requested metric."""
+        return self._metric_computation_handler.get_metric(metric=metric)
+
+    def get_metrics(
+        self,
+        metrics: Dict[str, MetricConfiguration],
+    ) -> Dict[str, Any]:
+        """
+        Convenience method that resolves requested metrics (specified as dictionary, keyed by MetricConfiguration ID).
+
+        Args:
+            metrics: Dictionary of desired metrics to be resolved; metric_name is key and MetricConfiguration is value.
+
+        Returns:
+            Return Dictionary with requested metrics resolved, with metric_name as key and computed metric as value.
+        """
+        return self._metric_computation_handler.get_metrics(metrics=metrics)
+
+    def compute_metrics(
+        self,
+        metric_configurations: List[MetricConfiguration],
+    ) -> Dict[Tuple[str, str, str], Any]:
+        """
+        Convenience method that computes requested metrics (specified as elements of "MetricConfiguration" list).
+
+        Args:
+            metric_configurations: List of desired MetricConfiguration objects to be resolved.
+
+        Returns:
+            Dictionary with requested metrics resolved, with unique metric ID as key and computed metric as value.
+        """
+        return self._metric_computation_handler.compute_metrics(
+            metric_configurations=metric_configurations
+        )
+
+    def columns(self, domain_kwargs: Optional[Dict[str, Any]] = None) -> List[str]:
+        """
+        Convenience method to obtain Batch columns.
+        """
+        return self._metric_computation_handler.columns(domain_kwargs=domain_kwargs)
+
+    def head(
+        self,
+        n_rows: int = 5,
+        domain_kwargs: Optional[Dict[str, Any]] = None,
+        fetch_all: bool = False,
+    ) -> pd.DataFrame:
+        """
+        Convenience method to obtain Batch first few rows.
+        """
+        return self._metric_computation_handler.head(
+            n_rows=n_rows, domain_kwargs=domain_kwargs, fetch_all=fetch_all
+        )
 
     def __dir__(self) -> List[str]:
         """
@@ -211,35 +304,6 @@ class Validator:
                     enable = progress_bars["metric_calculations"]
 
         return enable
-
-    @property
-    def execution_engine(self) -> ExecutionEngine:
-        """Returns the execution engine being used by the validator at the given time"""
-        return self._execution_engine
-
-    @property
-    def batch_cache(self) -> BatchCache:
-        return self._batch_cache
-
-    @property
-    def show_progress_bars(self) -> bool:
-        return self._show_progress_bars
-
-    @show_progress_bars.setter
-    def show_progress_bars(self, enable: bool) -> None:
-        self._show_progress_bars = enable
-
-    @property
-    def data_context(self) -> Optional["DataContext"]:  # noqa: F821
-        return self._data_context
-
-    @property
-    def expose_dataframe_methods(self) -> bool:
-        return self._expose_dataframe_methods
-
-    @expose_dataframe_methods.setter
-    def expose_dataframe_methods(self, value: bool) -> None:
-        self._expose_dataframe_methods = value
 
     def __getattr__(self, name):
         name = name.lower()
@@ -431,7 +495,7 @@ class Validator:
             profiler_result: RuleBasedProfilerResult = profiler.run(
                 variables=None,
                 rules=None,
-                batch_list=list(self._batch_cache.batch_list.values()),
+                batch_list=list(self.batches.values()),
                 batch_request=None,
                 recompute_existing_parameter_values=False,
                 reconciliation_directives=DEFAULT_RECONCILATION_DIRECTIVES,
@@ -777,115 +841,6 @@ class Validator:
             expectation for expectation in keys if expectation.startswith("expect_")
         ]
 
-    def compute_metrics(
-        self,
-        metric_configurations: List[MetricConfiguration],
-    ) -> Dict[Tuple[str, str, str], Any]:
-        """
-        Args:
-            metric_configurations: List of desired MetricConfiguration objects to be resolved.
-
-        Returns:
-            Dictionary with requested metrics resolved, with unique metric ID as key and computed metric as value.
-        """
-        graph = ValidationGraph()
-
-        metric_configuration: MetricConfiguration
-        for metric_configuration in metric_configurations:
-            provider_cls, _ = get_metric_provider(
-                metric_configuration.metric_name, self._execution_engine
-            )
-
-            self._get_default_domain_kwargs(
-                metric_provider_cls=provider_cls,
-                metric_configuration=metric_configuration,
-            )
-            self._get_default_value_kwargs(
-                metric_provider_cls=provider_cls,
-                metric_configuration=metric_configuration,
-            )
-
-            self.build_metric_dependency_graph(
-                graph=graph,
-                execution_engine=self._execution_engine,
-                metric_configuration=metric_configuration,
-            )
-
-        resolved_metrics: Dict[Tuple[str, str, str], Any] = {}
-
-        # updates graph with aborted metrics
-        aborted_metrics_info: Dict[
-            Tuple[str, str, str],
-            Dict[str, Union[MetricConfiguration, Set[ExceptionInfo], int]],
-        ] = self.resolve_validation_graph(
-            graph=graph,
-            metrics=resolved_metrics,
-        )
-
-        if aborted_metrics_info:
-            logger.warning(
-                f"Exceptions\n{str(aborted_metrics_info)}\noccurred while resolving metrics."
-            )
-
-        return resolved_metrics
-
-    def get_metrics(
-        self,
-        metrics: Dict[str, MetricConfiguration],
-    ) -> Dict[str, Any]:
-        """
-        Args:
-            metrics: Dictionary of desired metrics to be resolved, with metric_name as key and MetricConfiguration as value.
-
-        Returns:
-            Return Dictionary with requested metrics resolved, with metric_name as key and computed metric as value.
-        """
-        resolved_metrics: Dict[Tuple[str, str, str], Any] = self.compute_metrics(
-            metric_configurations=list(metrics.values()),
-        )
-
-        return {
-            metric_configuration.metric_name: resolved_metrics[metric_configuration.id]
-            for metric_configuration in metrics.values()
-        }
-
-    @staticmethod
-    def _get_default_domain_kwargs(
-        metric_provider_cls: "MetricProvider",  # noqa: F821
-        metric_configuration: MetricConfiguration,
-    ) -> None:
-        for key in metric_provider_cls.domain_keys:
-            if (
-                key not in metric_configuration.metric_domain_kwargs
-                and key in metric_provider_cls.default_kwarg_values
-            ):
-                metric_configuration.metric_domain_kwargs[
-                    key
-                ] = metric_provider_cls.default_kwarg_values[key]
-
-    @staticmethod
-    def _get_default_value_kwargs(
-        metric_provider_cls: "MetricProvider",  # noqa: F821
-        metric_configuration: MetricConfiguration,
-    ) -> None:
-        for key in metric_provider_cls.value_keys:
-            if (
-                key not in metric_configuration.metric_value_kwargs
-                and key in metric_provider_cls.default_kwarg_values
-            ):
-                metric_configuration.metric_value_kwargs[
-                    key
-                ] = metric_provider_cls.default_kwarg_values[key]
-
-    def get_metric(
-        self,
-        metric: MetricConfiguration,
-    ) -> Any:
-        """return the value of the requested metric."""
-        return self.get_metrics(
-            metrics={metric.metric_name: metric},
-        )[metric.metric_name]
-
     def graph_validate(
         self,
         configurations: List[ExpectationConfiguration],
@@ -1032,11 +987,9 @@ class Validator:
                 )
                 for metric_configuration in validation_dependencies.values():
                     graph = ValidationGraph()
-                    self.build_metric_dependency_graph(
+                    self._metric_computation_handler.build_metric_dependency_graph(
                         graph=graph,
-                        execution_engine=self._execution_engine,
                         metric_configuration=metric_configuration,
-                        configuration=evaluated_config,
                         runtime_configuration=runtime_configuration,
                     )
                     expectation_validation_graph.update(graph=graph)
@@ -1091,7 +1044,7 @@ class Validator:
         aborted_metrics_info: Dict[
             Tuple[str, str, str],
             Dict[str, Union[MetricConfiguration, Set[ExceptionInfo], int]],
-        ] = self.resolve_validation_graph(
+        ] = self._metric_computation_handler.resolve_validation_graph(
             graph=validation_graph,
             metrics=metrics,
             runtime_configuration=runtime_configuration,
@@ -1160,179 +1113,6 @@ class Validator:
             evrs.append(result)
 
         return evrs
-
-    def build_metric_dependency_graph(
-        self,
-        graph: ValidationGraph,
-        execution_engine: ExecutionEngine,
-        metric_configuration: MetricConfiguration,
-        configuration: Optional[ExpectationConfiguration] = None,
-        runtime_configuration: Optional[dict] = None,
-    ) -> None:
-        """Obtain domain and value keys for metrics and proceeds to add these metrics to the validation graph
-        until all metrics have been added."""
-
-        metric_impl = get_metric_provider(
-            metric_configuration.metric_name, execution_engine=execution_engine
-        )[0]
-        metric_dependencies = metric_impl.get_evaluation_dependencies(
-            metric=metric_configuration,
-            configuration=configuration,
-            execution_engine=execution_engine,
-            runtime_configuration=runtime_configuration,
-        )
-
-        if len(metric_dependencies) == 0:
-            graph.add(
-                MetricEdge(
-                    left=metric_configuration,
-                )
-            )
-        else:
-            metric_configuration.metric_dependencies = metric_dependencies
-            for metric_dependency in metric_dependencies.values():
-                # TODO: <Alex>In the future, provide a more robust cycle detection mechanism.</Alex>
-                if metric_dependency.id == metric_configuration.id:
-                    logger.warning(
-                        f"Metric {str(metric_configuration.id)} has created a circular dependency"
-                    )
-                    continue
-                graph.add(
-                    MetricEdge(
-                        left=metric_configuration,
-                        right=metric_dependency,
-                    )
-                )
-                self.build_metric_dependency_graph(
-                    graph=graph,
-                    execution_engine=execution_engine,
-                    metric_configuration=metric_dependency,
-                    configuration=configuration,
-                    runtime_configuration=runtime_configuration,
-                )
-
-    def resolve_validation_graph(  # noqa: C901 - complexity 16
-        self,
-        graph: ValidationGraph,
-        metrics: Dict[Tuple[str, str, str], Any],
-        runtime_configuration: Optional[dict] = None,
-        min_graph_edges_pbar_enable: int = 0,  # Set to low number (e.g., 3) to suppress progress bar for small graphs.
-    ) -> Dict[
-        Tuple[str, str, str],
-        Dict[str, Union[MetricConfiguration, Set[ExceptionInfo], int]],
-    ]:
-        if runtime_configuration is None:
-            runtime_configuration = {}
-
-        if runtime_configuration.get("catch_exceptions", True):
-            catch_exceptions = True
-        else:
-            catch_exceptions = False
-
-        failed_metric_info: Dict[
-            Tuple[str, str, str],
-            Dict[str, Union[MetricConfiguration, Set[ExceptionInfo], int]],
-        ] = {}
-        aborted_metrics_info: Dict[
-            Tuple[str, str, str],
-            Dict[str, Union[MetricConfiguration, Set[ExceptionInfo], int]],
-        ] = {}
-
-        ready_metrics: Set[MetricConfiguration]
-        needed_metrics: Set[MetricConfiguration]
-
-        exception_info: ExceptionInfo
-
-        progress_bar: Optional[tqdm] = None
-
-        done: bool = False
-        while not done:
-            ready_metrics, needed_metrics = self._parse_validation_graph(
-                validation_graph=graph, metrics=metrics
-            )
-
-            # Check to see if the user has disabled progress bars
-            disable = not self._show_progress_bars
-            if len(graph.edges) < min_graph_edges_pbar_enable:
-                disable = True
-
-            if progress_bar is None:
-                # noinspection PyProtectedMember,SpellCheckingInspection
-                progress_bar = tqdm(
-                    total=len(ready_metrics) + len(needed_metrics),
-                    desc="Calculating Metrics",
-                    disable=disable,
-                )
-            progress_bar.update(0)
-            progress_bar.refresh()
-
-            computable_metrics = set()
-
-            for metric in ready_metrics:
-                if (
-                    metric.id in failed_metric_info
-                    and failed_metric_info[metric.id]["num_failures"]
-                    >= MAX_METRIC_COMPUTATION_RETRIES
-                ):
-                    aborted_metrics_info[metric.id] = failed_metric_info[metric.id]
-                else:
-                    computable_metrics.add(metric)
-
-            try:
-                metrics.update(
-                    self._resolve_metrics(
-                        execution_engine=self._execution_engine,
-                        metrics_to_resolve=computable_metrics,
-                        metrics=metrics,
-                        runtime_configuration=runtime_configuration,
-                    )
-                )
-                progress_bar.update(len(computable_metrics))
-                progress_bar.refresh()
-            except MetricResolutionError as err:
-                if catch_exceptions:
-                    exception_traceback = traceback.format_exc()
-                    exception_message = str(err)
-                    exception_info = ExceptionInfo(
-                        exception_traceback=exception_traceback,
-                        exception_message=exception_message,
-                    )
-                    for failed_metric in err.failed_metrics:
-                        if failed_metric.id in failed_metric_info:
-                            failed_metric_info[failed_metric.id]["num_failures"] += 1
-                            failed_metric_info[failed_metric.id]["exception_info"].add(
-                                exception_info
-                            )
-                        else:
-                            failed_metric_info[failed_metric.id] = {}
-                            failed_metric_info[failed_metric.id][
-                                "metric_configuration"
-                            ] = failed_metric
-                            failed_metric_info[failed_metric.id]["num_failures"] = 1
-                            failed_metric_info[failed_metric.id]["exception_info"] = {
-                                exception_info
-                            }
-                else:
-                    raise err
-            except Exception as e:
-                if catch_exceptions:
-                    logger.error(
-                        f"""Caught exception {str(e)} while trying to resolve a set of {len(ready_metrics)} metrics; \
-aborting graph resolution.
-"""
-                    )
-                    done = True
-                else:
-                    raise e
-
-            if (len(ready_metrics) + len(needed_metrics) == 0) or (
-                len(ready_metrics) == len(aborted_metrics_info)
-            ):
-                done = True
-
-        progress_bar.close()
-
-        return aborted_metrics_info
 
     def append_expectation(self, expectation_config: ExpectationConfiguration) -> None:
         """This method is a thin wrapper for ExpectationSuite.append_expectation"""
@@ -2032,98 +1812,6 @@ aborting graph resolution.
 
         new_function = self.expectation(argspec)(function)
         return new_function(self, *args, **kwargs)
-
-    def columns(self, domain_kwargs: Optional[Dict[str, Any]] = None) -> List[str]:
-        if domain_kwargs is None:
-            domain_kwargs = {
-                "batch_id": self._execution_engine.batch_data_cache.active_batch_data_id,
-            }
-
-        columns: List[str] = self.get_metric(
-            metric=MetricConfiguration(
-                metric_name="table.columns",
-                metric_domain_kwargs=domain_kwargs,
-            )
-        )
-
-        return columns
-
-    def head(
-        self,
-        n_rows: int = 5,
-        domain_kwargs: Optional[Dict[str, Any]] = None,
-        fetch_all: bool = False,
-    ) -> pd.DataFrame:
-        if domain_kwargs is None:
-            domain_kwargs = {
-                "batch_id": self._execution_engine.batch_data_cache.active_batch_data_id,
-            }
-
-        data: Any = self.get_metric(
-            metric=MetricConfiguration(
-                metric_name="table.head",
-                metric_domain_kwargs=domain_kwargs,
-                metric_value_kwargs={
-                    "n_rows": n_rows,
-                    "fetch_all": fetch_all,
-                },
-            )
-        )
-
-        df: pd.DataFrame
-        if isinstance(
-            self._execution_engine, (PandasExecutionEngine, SqlAlchemyExecutionEngine)
-        ):
-            df = pd.DataFrame(data=data)
-        elif isinstance(self._execution_engine, SparkDFExecutionEngine):
-            rows: List[Dict[str, Any]] = [datum.asDict() for datum in data]
-            df = pd.DataFrame(data=rows)
-        else:
-            raise GreatExpectationsError(
-                "Unsupported or unknown ExecutionEngine type encountered in Validator class."
-            )
-
-        return df.reset_index(drop=True, inplace=False)
-
-    @staticmethod
-    def _parse_validation_graph(
-        validation_graph: ValidationGraph,
-        metrics: Dict[Tuple[str, str, str], Any],
-    ) -> Tuple[Set[MetricConfiguration], Set[MetricConfiguration]]:
-        """Given validation graph, returns the ready and needed metrics necessary for validation using a traversal of
-        validation graph (a graph structure of metric ids) edges"""
-        unmet_dependency_ids = set()
-        unmet_dependency = set()
-        maybe_ready_ids = set()
-        maybe_ready = set()
-
-        for edge in validation_graph.edges:
-            if edge.left.id not in metrics:
-                if edge.right is None or edge.right.id in metrics:
-                    if edge.left.id not in maybe_ready_ids:
-                        maybe_ready_ids.add(edge.left.id)
-                        maybe_ready.add(edge.left)
-                else:
-                    if edge.left.id not in unmet_dependency_ids:
-                        unmet_dependency_ids.add(edge.left.id)
-                        unmet_dependency.add(edge.left)
-
-        return maybe_ready - unmet_dependency, unmet_dependency
-
-    @staticmethod
-    def _resolve_metrics(
-        execution_engine: ExecutionEngine,
-        metrics_to_resolve: Iterable[MetricConfiguration],
-        metrics: Dict[Tuple[str, str, str], Any] = None,
-        runtime_configuration: dict = None,
-    ) -> Dict[Tuple[str, str, str], MetricConfiguration]:
-        """A means of accessing the Execution Engine's resolve_metrics method, where missing metric configurations are
-        resolved"""
-        return execution_engine.resolve_metrics(
-            metrics_to_resolve=metrics_to_resolve,
-            metrics=metrics,
-            runtime_configuration=runtime_configuration,
-        )
 
     def _initialize_expectations(
         self,
