+++ tests/data_context/datasource/test_data_context_datasource_non_sql_methods_sparkdf_execution_engine.py
@@ -104,9 +104,9 @@ def test_get_validator(context_with_single_titanic_csv_spark):
         batch_request=batch_request, expectation_suite_name="temp_suite"
     )
     assert isinstance(my_validator, Validator)
-    assert len(my_validator.batch_cache.batch_list) == 1
-    assert my_validator.batch_cache.active_batch.data.dataframe.count() == 1313
-    assert len(my_validator.batch_cache.active_batch.data.dataframe.columns) == 7
+    assert len(my_validator.batches) == 1
+    assert my_validator.active_batch.data.dataframe.count() == 1313
+    assert len(my_validator.active_batch.data.dataframe.columns) == 7
 
 
 @pytest.mark.integration
