+++ tests/data_context/datasource/test_data_context_datasource_runtime_data_connector_pandas_execution_engine.py
@@ -292,7 +292,7 @@ def test_batch_data_get_validator_successful_specification_pandas_engine_named_a
     )
     assert isinstance(my_validator, Validator)
     assert (
-        my_validator.batch_cache.active_batch.batch_definition.batch_identifiers
+        my_validator.active_batch.batch_definition.batch_identifiers
         == batch_identifiers
     )
 
@@ -651,6 +651,6 @@ def test_file_path_get_validator_successful_specification_pandas_engine_named_as
     )
     assert isinstance(my_validator, Validator)
     assert (
-        my_validator.batch_cache.active_batch.batch_definition.batch_identifiers
+        my_validator.active_batch.batch_definition.batch_identifiers
         == batch_identifiers
     )
