+++ tests/data_context/datasource/test_data_context_datasource_sql_methods.py
@@ -198,7 +198,7 @@ def test_get_validator(data_context_with_simple_sql_datasource_for_testing_get_b
                     data_connector_query=IDDict(batch_filter_parameters={}),
                 ),
                 expectation_suite_name="my_expectations",
-            ).batch_cache.batch_list
+            ).batches
         )
         == 34
     )
@@ -213,7 +213,7 @@ def test_get_validator(data_context_with_simple_sql_datasource_for_testing_get_b
                     data_asset_name="table_partitioned_by_date_column__A",
                 ),
                 expectation_suite_name="my_expectations",
-            ).batch_cache.batch_list
+            ).batches
         )
         == 34
     )
