+++ tests/integration/fixtures/yellow_tripdata_pandas_fixture/multiple_batch_requests_one_validator_one_step.py
@@ -40,12 +40,12 @@ validator: Validator = context.get_validator(
     batch_request_list=[jan_batch_request, feb_batch_request, march_batch_request],
     expectation_suite=suite,
 )
-assert validator.batch_cache.active_batch_definition.batch_identifiers["month"] == "03"
-assert validator.batch_cache.active_batch_definition.batch_identifiers["year"] == "2019"
+assert validator.active_batch.batch_definition.batch_identifiers["month"] == "03"
+assert validator.active_batch.batch_definition.batch_identifiers["year"] == "2019"
 
 # Get the list of all batches contained by the Validator for use in the BatchFileter
 total_batch_definition_list: List = [
-    v.batch_definition for k, v in validator.batch_cache.batch_list.items()
+    v.batch_definition for k, v in validator.batches.items()
 ]
 
 # Filter to all batch_definitions prior to March
