+++ tests/integration/fixtures/yellow_tripdata_pandas_fixture/one_multi_batch_request_one_validator.py
@@ -26,13 +26,13 @@ validator_multi_batch = context.get_validator(
 
 # The active batch should be December, as this should be the last one loaded. Confirming here.
 assert (
-    validator_multi_batch.batch_cache.active_batch_definition.batch_identifiers["month"]
+    validator_multi_batch.active_batch.batch_definition.batch_identifiers["month"]
     == "12"
 )
 
 # Get the list of all batches contained by the Validator for use in the BatchFilter
 total_batch_definition_list: list = [
-    v.batch_definition for k, v in validator_multi_batch.batch_cache.batch_list.items()
+    v.batch_definition for k, v in validator_multi_batch.batches.items()
 ]
 
 # Filter to all batch_definitions prior to December
